Add export instances toggle for scene cache
