+++ Plugins~/Src/MeshSyncClientBlender/CMakeLists.txt
@@ -75,6 +75,7 @@ install(FILES
     python/unity_mesh_sync_common.py
     python/unity_mesh_sync_installation.py
     python/unity_mesh_sync_preferences.py
+    python/unity_mesh_sync_scene_cache.py
     DESTINATION ${PYD_DIR})
 install(TARGETS ${MESHSYNC_CLIENT_BLENDER} DESTINATION ${PYD_DIR})
